feat: Implements suggestions about testing
The suggestion on this review was to improve the tests. One was to use env vars rather than hardcoded values. The other one was to remove the pagination with all the fetched data but rather load the data from the DB as the request ask for them. The latter will be implemented in a new commit

Resolves: #1 (Testing)